pyproject.toml
[project]
name = "django-rest-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.0.6",
    "djangorestframework==3.15.1",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.3.0",
    "pip>=25.1.1",
    "requests>=2.32.4",
]
